+++ simfin/__init__.py
@@ -1,5 +1,5 @@
 # This is also defined in setup.py and must be updated in both places.
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 # Expose the following as top-level imports.
 
@@ -10,7 +10,8 @@ from simfin.bulk import load_balance, load_balance_banks, load_balance_insurance
 from simfin.bulk import load_cashflow, load_cashflow_banks, load_cashflow_insurance
 
 from simfin.utils import add_date_offset
-from simfin.resample import resample, resample_daily
+from simfin.resample import resample, resample_daily, asfreq, asfreq_daily
+from simfin.resample import reindex, index_union
 
 from simfin.config import set_api_key, load_api_key, get_api_key
 from simfin.config import set_data_dir, get_data_dir, get_download_dir